Refactor auth0 management a bit.
