+++ src/backend/aspen/cli/sync_auth0.py
@@ -1,16 +1,10 @@
 #!/usr/bin/env python3
 import logging
-import os
-import random
-import string
-from functools import cache, partial
-from typing import Any, Callable, List, MutableSequence, Tuple, TypedDict
 
 import click
 import sqlalchemy as sa
-from auth0.v3 import authentication as auth0_authentication
-from auth0.v3.management import Auth0
 from sqlalchemy.orm.session import Session
+from aspen.auth.auth0_management import Auth0Client, Auth0Org, Auth0User
 
 from aspen.config.config import Config
 from aspen.database.connection import (
@@ -19,124 +13,8 @@ from aspen.database.connection import (
     session_scope,
     SqlAlchemyInterface,
 )
+from aspen.config.config import Config
 from aspen.database.models import Group, User
-
-
-class Auth0Org(TypedDict):
-    id: str
-    name: str
-    display_name: str
-
-
-class Auth0User(TypedDict):
-    user_id: str
-    name: str
-    email: str
-
-
-class Auth0Role(TypedDict):
-    id: str
-    name: str
-
-
-def generate_password(length: int = 22) -> str:
-    possible_characters = (
-        string.ascii_uppercase + string.ascii_lowercase + string.digits
-    )
-    return "".join(random.choice(possible_characters) for _ in range(length))
-
-
-class Auth0Client:
-    def __init__(self) -> None:
-        # TODO these will need to be read from settings instead of env.
-        client_id: str = os.environ["AUTH0_CLIENT_ID"]
-        client_secret: str = os.environ["AUTH0_CLIENT_SECRET"]
-        domain: str = os.environ["AUTH0_DOMAIN"]
-        auth_req = auth0_authentication.GetToken(domain)
-        token = auth_req.client_credentials(
-            client_id, client_secret, f"https://{domain}/api/v2/"
-        )
-
-        self.client: Auth0 = Auth0(domain=domain, token=token["access_token"])
-
-    def get_all_results(self, endpoint: Callable, key: str) -> List[Any]:
-        # Auth0 paginates results. We don't have a crazy amount of data in auth0 so we can
-        # afford to just paginate through all the results and hold everything in memory.
-        results = []
-        page = 0
-        per_page = 25
-        while True:
-            resp = endpoint(page=page, per_page=per_page)
-            last_result = resp["start"] + resp["limit"]
-            results.extend(resp[key])
-            if last_result >= resp["total"]:
-                return results
-            page += 1
-
-    def get_users(self) -> List[Auth0User]:
-        return self.get_all_results(self.client.users.list, "users")
-
-    def get_orgs(self) -> List[Auth0Org]:
-        return self.get_all_results(
-            self.client.organizations.all_organizations, "organizations"
-        )
-
-    @cache
-    def get_auth0_role(self, role_name: str) -> Auth0Role:
-        all_roles = self.get_roles()
-        for role in all_roles:
-            if role["name"] == role_name:
-                return role
-        raise Exception("Role not found")
-
-    @cache
-    def get_roles(self) -> List[Auth0Role]:
-        return self.get_all_results(self.client.roles.list, "roles")
-
-    def get_org_members(self, org: Auth0Org) -> List[Auth0User]:
-        return self.get_all_results(
-            partial(self.client.organizations.all_organization_members, org["id"]),
-            "members",
-        )
-
-    def add_org_member(self, org: Auth0Org, user_id: str) -> None:
-        self.client.organizations.create_organization_members(
-            org["id"], {"members": [user_id]}
-        )
-        member_role = self.get_auth0_role("member")
-        self.client.organizations.create_organization_member_roles(
-            org["id"], user_id, {"roles": [member_role["id"]]}
-        )
-
-    def remove_org_member(self, org: Auth0Org, user_id: str) -> None:
-        self.client.organizations.delete_organization_members(
-            org["id"], {"members": [user_id]}
-        )
-
-    def add_org(self, group_id: int, org_name: str) -> None:
-        body = {
-            "name": f"group-{group_id}",
-            "display_name": org_name,
-        }
-        self.client.organizations.create_organization(body)
-
-    def delete_org(self, org_id: str) -> None:
-        self.client.organizations.delete_organization(org_id)
-
-    def create_user(self, email: str, name: str) -> None:
-        body = {
-            "email": email,
-            "user_metadata": {},
-            "email_verified": True,
-            "name": name,
-            "verify_email": False,
-            "password": generate_password(),
-            "connection": "Username-Password-Authentication",
-        }
-        self.client.users.create(body)
-
-    def delete_user(self, auth0_user_id: str) -> None:
-        self.client.users.delete(auth0_user_id)
 
 
 def find_missing_objects(
@@ -354,10 +232,10 @@ class SuperSyncer:
         )
         for org in auth0_orgs:
             auth0_memberships: List[Auth0User] = self.auth0_client.get_org_members(org)
-            # TODO, we might want to stuff the auth0 group IDs into the groups table to
+            # TODO, we might want to stuff the auth0 group ID's into the groups table to
             # make this simpler.
             found_groups: MutableSequence[Group] = [
-                group for group in db_groups if group.name == org["display_name"]
+                group for group in found_groups if group.name == org["display_name"]
             ]
             if not found_groups:
                 # We're assuming that at this point in the script, we've already sync'd
@@ -425,7 +303,13 @@ def cli(
     sync_users: bool,
     sync_memberships: bool,
 ) -> None:
-    auth0_client = Auth0Client()
+    # TODO - pass in client id, secret, domain here!!!
+
+    config = Config()
+    client_id: str = config.AUTH0_MANAGEMENT_CLIENT_ID
+    client_secret: str = config.AUTH0_MANAGEMENT_CLIENT_SECRET
+    domain: str = config.AUTH0_MANAGEMENT_DOMAIN
+    auth0_client = Auth0Client(client_id, client_secret, domain)
 
     logging.basicConfig(
         format="%(levelname)s %(asctime)s - %(message)s", level=logging.INFO
